requirements.txt
alembic==1.18.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Deprecated==1.3.1
fastapi==0.109.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
kombu==5.6.2
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==8.2.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
protobuf==4.21.12
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
redis==6.4.0
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.45
starlette==0.35.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.27.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wrapt==2.0.1
